refactor: 카드 모달창 경로 변경
#69
