+++ src/components/CardModal/Content/Description.styles.ts
