+++ src/components/CardModal/Content/styles.ts
