+++ src/components/CardModal/Sidebar/Sidebar.styles.ts
