+++ src/components/CardModal/styles.ts
